requirements.txt
websocket-client==1.6.4
python-dotenv==1.0.0
Flask==3.0.0
ccxt==4.1.87
logging==0.5.1.2 